Breakout development services into their own project and remove duplicate/redundant properties
