+++ Demo/Controller.cs
@@ -21,7 +21,7 @@ public class MyController : Controller
 {
     private IFido2 _fido2;
     public static IMetadataService _mds;
-    public static readonly DevelopmentInMemoryStore DemoStorage = new DevelopmentInMemoryStore();
+    public static readonly DevelopmentInMemoryStore DemoStorage = new();
 
     public MyController(IFido2 fido2)
     {
@@ -119,7 +119,6 @@ public class MyController : Controller
             // 3. Store the credentials in db
             DemoStorage.AddCredentialToUser(options.User, new StoredCredential
             {
-                Type = success.Result.Type,
                 Id = success.Result.Id,
                 Descriptor = new PublicKeyCredentialDescriptor(success.Result.Id),
                 PublicKey = success.Result.PublicKey,
@@ -129,8 +128,8 @@ public class MyController : Controller
                 RegDate = DateTime.Now,
                 AaGuid = success.Result.AaGuid,
                 Transports = success.Result.Transports,
-                BE = success.Result.BE,
-                BS = success.Result.BS,
+                IsBackupEligible = success.Result.BE,
+                IsBackedUp = success.Result.BS,
                 AttestationObject = success.Result.AttestationObject,
                 AttestationClientDataJSON = success.Result.AttestationClientDataJSON,
                 DevicePublicKeys = new List<byte[]>() { success.Result.DevicePublicKey }
@@ -204,9 +203,9 @@ public class MyController : Controller
             var creds = DemoStorage.GetCredentialById(clientResponse.Id) ?? throw new Exception("Unknown credentials");
 
             // 3. Get credential counter from database
-            var storedCounter = creds.SignatureCounter;
+            var storedCounter = creds.SignCount;
 
-            // 4. Create callback to check if userhandle owns the credentialId
+            // 4. Create callback to check if the user handle owns the credentialId
             IsUserHandleOwnerOfCredentialIdAsync callback = static async (args, cancellationToken) =>
             {
                 var storedCreds = await DemoStorage.GetCredentialsByUserHandleAsync(args.UserHandle, cancellationToken);
