+++ Demo/Pages/dashboard.cshtml
@@ -62,7 +62,7 @@
                 <tr>
                     <td>@item.RegDate</td>
                     <td><img src="@icon" /> @desc</td>
-                    <td>@item.SignatureCounter</td>
+                    <td>@item.SignCount</td>
                     <td>@item.AaGuid</td>
                     <td>@userHandle</td>
                     <td>@userId</td>
