+++ fido2-net-lib.sln
@@ -30,6 +30,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlazorWasmDemo.Server", "Bl
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlazorWasmDemo.Client", "BlazorWasmDemo\Client\BlazorWasmDemo.Client.csproj", "{EAFF4332-145B-4719-B261-827CEB447A62}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Fido2.Development", "Src\Fido2.Development\Fido2.Development.csproj", "{08089F2C-A0EE-4B2A-BFE6-B959DEAA4971}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -64,6 +66,10 @@ Global
 		{BB3D1D09-1506-4DF9-8211-8B284E12B0F8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BB3D1D09-1506-4DF9-8211-8B284E12B0F8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BB3D1D09-1506-4DF9-8211-8B284E12B0F8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F314C16B-3E3D-40E6-9B4F-36142B739478}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F314C16B-3E3D-40E6-9B4F-36142B739478}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F314C16B-3E3D-40E6-9B4F-36142B739478}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F314C16B-3E3D-40E6-9B4F-36142B739478}.Release|Any CPU.Build.0 = Release|Any CPU
 		{0EC19932-290A-4D92-9603-B0533DD022D9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{0EC19932-290A-4D92-9603-B0533DD022D9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0EC19932-290A-4D92-9603-B0533DD022D9}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -72,10 +78,10 @@ Global
 		{EAFF4332-145B-4719-B261-827CEB447A62}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EAFF4332-145B-4719-B261-827CEB447A62}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EAFF4332-145B-4719-B261-827CEB447A62}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F314C16B-3E3D-40E6-9B4F-36142B739478}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F314C16B-3E3D-40E6-9B4F-36142B739478}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F314C16B-3E3D-40E6-9B4F-36142B739478}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F314C16B-3E3D-40E6-9B4F-36142B739478}.Release|Any CPU.Build.0 = Release|Any CPU
+		{08089F2C-A0EE-4B2A-BFE6-B959DEAA4971}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{08089F2C-A0EE-4B2A-BFE6-B959DEAA4971}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{08089F2C-A0EE-4B2A-BFE6-B959DEAA4971}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{08089F2C-A0EE-4B2A-BFE6-B959DEAA4971}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
